Add logic to read and cache current values from device
When opening a device, query the current state of the input elements. For
KEY, LED, SW and SND this is done by using the appropriate macros which
return the state for all the elements at once, in one large bitmap.

For ABS, we have to get the values through AbsInfo individually.

Store the current value in an exported hashmap.